fitness-tracker-app/fitness-tracker/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
drf-yasg==1.21.7
gunicorn==21.2.0
whitenoise==6.5.0
dj-database-url==2.1.0
</antArt
